requirements.txt
asgiref==3.8.1
Django==4.2.15
djangorestframework==3.15.2
numpy==2.0.2
psycopg==3.2.1
psycopg2==2.9.9
pydicom==2.4.4
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
